优化代码2

+++ kernel/include/int/intTable.h
@@ -5,7 +5,7 @@ typedef struct {
     unsigned int vector;
     unsigned int errCode;
     unsigned int eip;
-    unsigned short cs;
+    unsigned int cs;
     unsigned int eflags;
 } IsrParam;
 #pragma pack(pop)
